backend/requirements.txt
# Flask and web framework
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# SpoonOS agent framework
# Note: Install spoonos from appropriate source
# pip install spoonos or follow SpoonOS installation instructions

# Neo blockchain integration
# Note: Install spoon-toolkit for Neo blockchain operations
# pip install spoon-toolkit or follow spoon-toolkit installation instructions

# Web3 and blockchain
web3==6.15.1

# Text-to-speech
elevenlabs==1.0.0

# Utilities
requests==2.31.0

# Development dependencies (optional)
# pytest==7.4.3
# pytest-flask==1.3.0
# black==23.11.0
# flake8==6.1.0

